+++ CHANGELOG.md
@@ -7,6 +7,7 @@
         feed DOM page snapshots and new paths back to the `Framework`.
 - Added `State` -- Stores and provides access to the system's state.
     - `Issues` -- Stores logged `Issue` objects.
+    - `Audit` -- Stores audit operations.
 - Removed the `Spider`.
     - The Framework has grown to encompass a process providing the same
         functionality as a result of `Browser` analysis.